+++ config/react-ts/webpack.dev.js
@@ -1,9 +1,3 @@
-const { webpackMerge } = require('../base/util')
 const webpackBase = require('../base/webpack.dev')
 
-const init = (config, iEnv) => {
-  return webpackMerge(
-    webpackBase(config, iEnv)
-  )
-}
-module.exports = init
+module.exports = webpackBase